--- eagleye_util/nmea2fix/package.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>nmea2fix</name>
-  <version>1.0.0</version>
-  <description>nmea2fix package</description>
-
-  <maintainer email="osamu.sekino@tier4.jp">OsamuSekino</maintainer>
-
-  <license>BSD3</license>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>rospy</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>eagleye_msgs</build_depend>
-  <build_depend>rtklib_msgs</build_depend>
-  <build_depend>tf</build_depend>
-  <build_depend>coordinate</build_depend>
-  <build_export_depend>roscpp</build_export_depend>
-  <build_export_depend>rospy</build_export_depend>
-  <build_export_depend>std_msgs</build_export_depend>
-  <build_export_depend>eagleye_msgs</build_export_depend>
-  <build_export_depend>rtklib_msgs</build_export_depend>
-  <build_export_depend>tf</build_export_depend>
-  <build_export_depend>coordinate</build_export_depend>
-  <exec_depend>roscpp</exec_depend>
-  <exec_depend>rospy</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
-  <exec_depend>eagleye_msgs</exec_depend>
-  <exec_depend>rtklib_msgs</exec_depend>
-  <exec_depend>tf</exec_depend>
-  <exec_depend>coordinate</exec_depend>
-
-  <export>
-
-  </export>
-</package>
